dag 16 b
- eerst de combinatie van vermenigvuldiging en som voorkomen door aan het einde
van de loop niet te gaan kijken of er nog meer enen aanwezig waren
- zorgen dat 2021 niet 1013 oplevert doordat er ergens nog met een Integer gewerkt werd
- berekening omzetten van Integer naar Long

+++ src/net/gideonse/aoc2021_16/GroterDanPakket.java
@@ -8,15 +8,15 @@ public class GroterDanPakket extends AbstractOperatorPakket {
     }
 
     @Override
-    public Integer doeBerekening(List<? extends Pakket> kinderen) {
-        final List<Integer> kleinkinderen = kinderen.stream()
+    public Long doeBerekening(List<? extends Pakket> kinderen) {
+        final List<Long> kleinkinderen = kinderen.stream()
                 .map(kind -> kind.doeBerekening(kind.getKinderen())).toList();
         assert kleinkinderen.size() == 2;
 
         System.out.print("GroterDanPakket.doeBerekening");
         System.out.println(" kleinkinderen = " + kleinkinderen);
 
-        return kleinkinderen.get(0) > kleinkinderen.get(1) ? 1 : 0;
+        return kleinkinderen.get(0) > kleinkinderen.get(1) ? 1L : 0L;
     }
 
     @Override
